requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.4; python_version >= '3.6'
django==3.2.4
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4; python_version >= '3.5'
